refactor: move files to the controller folder to follow an mvc pattern
