+++ controllers/teste3.js
@@ -1,5 +1,5 @@
-const readFile = require('./utils/readFile');
-const writeFile = require('./utils/writeFile');
+const readFile = require('../utils/readFile');
+const writeFile = require('../utils/writeFile');
 
 const data = readFile('./fakeData.json');
 const deleteUser = (req, res) => {
