+++ fakeData.json
@@ -1,1 +1,1 @@
-[{"id":"1","name":"João Oliveira","job":"Desenvolvedor","counter":1}]
+[{"id":1,"name":"João Oliveira","job":"Desenvolvedor"}]
